requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
python-dotenv>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
